+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Core: updated Japanese holidays
+
 25-05-2023: 6.6.162
 - Email / OAuth2Client: More friendly message when refresh token needs to be renewed
 - Addressbook: generate search words for contact urls too